# sql: store information about keys in orderingInfo, elide distinctNode