--- a/pkg/sql/logictest/testdata/logic_test/aggregate
+++ b/pkg/sql/logictest/testdata/logic_test/aggregate
@@ -259,7 +259,7 @@ EXPLAIN (VERBOSE) SELECT COUNT(*), k+v FROM kv GROUP BY k+v
 0  ·       group by     @1-@1                  ·                               ·
 1  render  ·            ·                      ("k + v")                       ·
 1  ·       render 0     test.kv.k + test.kv.v  ·                               ·
-2  scan    ·            ·                      (k, v, w[omitted], s[omitted])  ·
+2  scan    ·            ·                      (k, v, w[omitted], s[omitted])  key(k)
 2  ·       table        kv@primary             ·                               ·
 2  ·       spans        ALL                    ·                               ·
 
@@ -286,10 +286,10 @@ EXPLAIN (VERBOSE) SELECT COUNT(*), k+v FROM kv GROUP BY k, v
 1  ·       aggregate 1  test.kv.k     ·                               ·
 1  ·       aggregate 2  test.kv.v     ·                               ·
 1  ·       group by     @1-@2         ·                               ·
-2  render  ·            ·             (k, v)                          ·
+2  render  ·            ·             (k, v)                          key(k)
 2  ·       render 0     test.kv.k     ·                               ·
 2  ·       render 1     test.kv.v     ·                               ·
-3  scan    ·            ·             (k, v, w[omitted], s[omitted])  ·
+3  scan    ·            ·             (k, v, w[omitted], s[omitted])  key(k)
 3  ·       table        kv@primary    ·                               ·
 3  ·       spans        ALL           ·                               ·
 
@@ -1153,7 +1153,7 @@ EXPLAIN (TYPES) SELECT COUNT(*) FILTER (WHERE k > 5) FROM filter_test GROUP BY v
 1  render  ·            ·                                                               (v int, "k > 5" bool)                                          ·
 1  ·       render 0     (v)[int]                                                        ·                                                              ·
 1  ·       render 1     ((k)[int] > (5)[int])[bool]                                     ·                                                              ·
-2  scan    ·            ·                                                               (k int, v int, mark[omitted] bool, rowid[hidden,omitted] int)  ·
+2  scan    ·            ·                                                               (k int, v int, mark[omitted] bool, rowid[hidden,omitted] int)  key(rowid)
 2  ·       table        filter_test@primary                                             ·                                                              ·
 2  ·       spans        ALL                                                             ·                                                              ·
 
@@ -1225,7 +1225,7 @@ EXPLAIN (VERBOSE) SELECT MIN(v) FROM opt_test
 0  ·       aggregate 0  min(test.opt_test.v)  ·                ·
 1  render  ·            ·                     (v)              +v
 1  ·       render 0     test.opt_test.v       ·                ·
-2  scan    ·            ·                     (k[omitted], v)  +v
+2  scan    ·            ·                     (k[omitted], v)  key(k,v); +v
 2  ·       table        opt_test@v            ·                ·
 2  ·       spans        /#-                   ·                ·
 2  ·       limit        1                     ·                ·
@@ -1250,7 +1250,7 @@ EXPLAIN (VERBOSE) SELECT MIN(v) FROM opt_test WHERE k <> 4
 0  ·       aggregate 0  min(test.opt_test.v)  ·           ·
 1  render  ·            ·                     (v)         +v
 1  ·       render 0     test.opt_test.v       ·           ·
-2  scan    ·            ·                     (k, v)      +v
+2  scan    ·            ·                     (k, v)      key(k,v); +v
 2  ·       table        opt_test@v            ·           ·
 2  ·       spans        /#-                   ·           ·
 2  ·       filter       k != 4                ·           ·
@@ -1270,7 +1270,7 @@ EXPLAIN (VERBOSE) SELECT MIN(v+1) FROM opt_test WHERE k <> 4
 0  ·       aggregate 0  min(test.opt_test.v + 1)            ·               ·
 1  render  ·            ·                                   ("v + 1")       ·
 1  ·       render 0     test.opt_test.v + 1                 ·               ·
-2  scan    ·            ·                                   (k, v)          ·
+2  scan    ·            ·                                   (k, v)          key(k)
 2  ·       table        opt_test@primary                    ·               ·
 2  ·       spans        /#-                                 ·               ·
 2  ·       filter       (k != 4) AND ((v + 1) IS NOT NULL)  ·               ·
@@ -1282,7 +1282,7 @@ EXPLAIN (VERBOSE) SELECT MIN(v) FROM opt_test GROUP BY k
 0  group  ·            ·                     ("min(v)")  ·
 0  ·      aggregate 0  min(test.opt_test.v)  ·           ·
 0  ·      group by     @1-@1                 ·           ·
-1  scan   ·            ·                     (k, v)      ·
+1  scan   ·            ·                     (k, v)      key(k)
 1  ·      table        opt_test@primary      ·           ·
 1  ·      spans        ALL                   ·           ·