--- a/pkg/sql/logictest/testdata/logic_test/alter_table
+++ b/pkg/sql/logictest/testdata/logic_test/alter_table
@@ -638,9 +638,9 @@ EXPLAIN (METADATA) ALTER TABLE s SPLIT AT SELECT k1,k2 FROM s ORDER BY k1 LIMIT 
 ----
 Level  Type    Field  Description  Columns               Ordering
 0      split   ·      ·            (key, pretty)         ·
-1      limit   ·      ·            (k1, k2)              +k1
-2      render  ·      ·            (k1, k2)              +k1
-3      scan    ·      ·            (k1, k2, v[omitted])  +k1
+1      limit   ·      ·            (k1, k2)              key(k1,k2); +k1
+2      render  ·      ·            (k1, k2)              key(k1,k2); +k1
+3      scan    ·      ·            (k1, k2, v[omitted])  key(k1,k2); +k1
 3      ·       table  s@primary    ·                     ·
 3      ·       spans  ALL          ·                     ·
 3      ·       limit  3            ·                     ·