--- a/pkg/util/fast_int_set.go
+++ b/pkg/util/fast_int_set.go
@@ -208,7 +208,7 @@ func (s *FastIntSet) SubsetOf(rhs FastIntSet) bool {
 	return (s1 & s2) == s1
 }
 
-func (s *FastIntSet) String() string {
+func (s FastIntSet) String() string {
 	var buf bytes.Buffer
 	buf.WriteByte('(')
 	first := true