… pkg/sql/bench_test.go
… pkg/sql/logictest/testdata/logic_test/join
… pkg/sql/logictest/testdata/logic_test/needed_columns
… pkg/sql/logictest/testdata/logic_test/order_by
… pkg/sql/logictest/testdata/logic_test/order_by_index
… pkg/sql/logictest/testdata/logic_test/ordinal_references
… pkg/sql/logictest/testdata/logic_test/ordinality
… pkg/sql/logictest/testdata/logic_test/ranges
… pkg/sql/logictest/testdata/logic_test/select_index
… pkg/sql/logictest/testdata/logic_test/select_non_covering_index
… pkg/sql/logictest/testdata/logic_test/select_non_covering_index_filtering
… pkg/sql/logictest/testdata/logic_test/stars_in_subexpressions
… pkg/sql/logictest/testdata/logic_test/subquery
… pkg/sql/logictest/testdata/logic_test/update
… pkg/sql/logictest/testdata/logic_test/window
… pkg/sql/ordering.go
… pkg/sql/ordering_test.go
… pkg/sql/ordinality.go
… pkg/sql/scan.go