#### sql: store information about keys in orderingInfo

We are now able to store information about column sets that form "keys", i.e.
all rows are distinct when projected on that set.

The existing structure had a very limited form of this, only keeping track if
the columns in the ordering form a key. Allowing arbitrary sets will enable more
optimizations.

#### sql: elide distinctNode if we have a key

An immediate application of the new key information is to elide distinct nodes
when possible.