@knz 2017-09-15
I'll review this once the first two PRs are merged, so that the diff is more manageable.