@RaduBerinde 2017-10-03
Rebased, ready to review.