@knz 2017-10-05
<img class="emoji" title=":lgtm_strong:" alt=":lgtm_strong:" align="absmiddle" src="https://reviewable.io/lgtm_strong.png" height="20" width="61"/> 

As a matter of principle, can you give an estimate in the commit message for the 1st commit of the added allocation overhead for typical (simple) queries? 

---

Reviewed 26 of 26 files at r1, 3 of 4 files at r2.
Review status: 27 of 28 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/18508#-:-KvgppfsvrJCM18zQYsr:b-i0141x)*
<!-- Sent from Reviewable.io -->